Add files via upload
In this exam, you are asked to calculate the most efficient way of dividing the land given the dimensions of the land X and Y, and the dimensions of the plots in a 2D array of booleans, then return the m^2 of unusable land by completing the divide_land() function defined below. 


int divide_land(int X, int Y, bool** possible_plots);

 



X, Y: dimensions X x Y of the total rectangular land of the island
possible_plots: a 2D array of booleans, where each value stands for the existence of a possible plot with the indices of the array cell as dimensions (for example, if possible_plot[2][3] == true, then 2 x 3 is a possible plot dimension that can be used)
to return: m^2 of unusable land, meaning the land that does not belong to any plot after all the dividing operations.

Constraints and Hints:

X and Y are integers <= 600
possible_plots is a 2D array with dimensions (X+1)x(Y+1). So, the maximum dimensions are 601x601. At any given time, there will be at most 200 "true" cells in the 2D array - but the complexity of the optimal solution does not depend on this value.
You are expected to use a dynamic programming approach to reduce complexity.